[pbernier] - Working on ft_strjoin_clean. ;)

+++ lib/libft/srcs/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: pbernier <pbernier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/09 00:54:25 by pbernier          #+#    #+#             */
-/*   Updated: 2017/08/18 13:13:55 by pbernier         ###   ########.fr       */
+/*   Updated: 2017/09/12 19:54:06 by pbernier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t		ft_strlen(const char *s)
 {
-	int		i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
