+++ lib/libft/srcs/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: pbernier <pbernier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/10 01:32:26 by pbernier          #+#    #+#             */
-/*   Updated: 2017/08/18 13:14:21 by pbernier         ###   ########.fr       */
+/*   Updated: 2017/09/12 20:19:57 by pbernier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int		i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] != '\0' || s2[i] != '\0')
